cart.js and its members, methods moved to module
