+++ amazon.html
@@ -55,7 +55,6 @@
       <div class="products-grid js-products-grid"></div>
     </div>
 
-  <script src="data/cart.js"></script>
   <script type="module" src="scripts/amazon.js";></script>  
   </body>
 </html>
